+++ picopt/handlers/png.py
@@ -5,6 +5,7 @@ from types import MappingProxyType
 from PIL.PngImagePlugin import PngImageFile
 from termcolor import cprint
 
+from picopt.handlers.convertible import CONVERTABLE_FORMAT_STRS, GIF_FORMAT_STR
 from picopt.handlers.handler import FileFormat
 from picopt.handlers.image import ImageHandler
 from picopt.pillow.png_bit_depth import png_bit_depth
@@ -17,6 +18,7 @@ class Png(ImageHandler):
     OUTPUT_FORMAT_STR = PngImageFile.format
     OUTPUT_FILE_FORMAT = FileFormat(OUTPUT_FORMAT_STR, True, False)
     INPUT_FILE_FORMATS = frozenset({OUTPUT_FILE_FORMAT})
+    CONVERT_FROM_FORMAT_STRS = frozenset(CONVERTABLE_FORMAT_STRS | {GIF_FORMAT_STR})
     PIL2_ARGS: MappingProxyType[str, bool] = MappingProxyType({"optimize": True})
     PROGRAMS: MappingProxyType[
         str, str | tuple[str, ...] | None